pyproject.toml
[tool.poetry]
name = "gclinic"
version = "0.1.0"
description = ""
authors = ["Muhammed Zeba <parice02@hotmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-environ = "^0.10.0"
django-easy-audit = "*"
django-formset = "^1.4"
django-import-export = "^3.3.1"
django-feather = "^0.3.0"
django-csp = "^3.7"
django-permissions-policy = "^4.18.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
django-model-utils = "^4.3.1"
django-autoslug = "^1.9.9"
pillow = "^10.2.0"
django-filter = "^24.2"
django-chartjs = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
